Slight fixes

+++ src/Main.java
@@ -40,7 +40,7 @@ public class Main {
                     String cutie_v = sc.next();
                     String tractiune = sc.next();
                     int masa = sc.nextInt();
-                    int val = sc.nextInt();
+                    float val = sc.nextFloat();
 
                     Sedan sedan_curent = new Sedan(an_fab, brand, model, culoare, motor, cutie_v, putere, tractiune, masa, val);
                     parcare.add(sedan_curent);
@@ -60,7 +60,7 @@ public class Main {
                     String cutie_v = sc.next();
                     int masa_rem = sc.nextInt();
                     int masa = sc.nextInt();
-                    int val = sc.nextInt();
+                    float val = sc.nextFloat();
 
                     TIR tir_curent = new TIR(an_fab, brand, model, culoare, motor, cutie_v, putere, masa_rem, masa, val);
                     parcare.add(tir_curent);
@@ -80,7 +80,7 @@ public class Main {
                     String cutie_v = sc.next();
                     String tip_teren = sc.next();
                     int masa = sc.nextInt();
-                    int val = sc.nextInt();
+                    float val = sc.nextFloat();
 
                     ATV atv_curent = new ATV(an_fab, brand, model, culoare, motor, cutie_v, putere, tip_teren, masa, val);
                     parcare.add(atv_curent);
@@ -100,7 +100,7 @@ public class Main {
                     String cutie_v = sc.next();
                     String clasa = sc.next();
                     int masa = sc.nextInt();
-                    int val = sc.nextInt();
+                    float val = sc.nextFloat();
 
                     Motocicleta moto_curent = new Motocicleta(an_fab, brand, model, culoare, motor, cutie_v, putere, clasa, masa, val);
                     parcare.add(moto_curent);
